+++ src/key.rs
@@ -3,13 +3,13 @@ use core::ops::Deref;
 use num_bigint::traits::ModInverse;
 use num_bigint::Sign::Plus;
 use num_bigint::{BigInt, BigUint};
-use num_traits::{One, ToPrimitive};
+use num_traits::{FromPrimitive, One, ToPrimitive};
 use rand_core::CryptoRngCore;
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 use zeroize::Zeroize;
 
-use crate::algorithms::{generate_multi_prime_key, generate_multi_prime_key_with_exp};
+use crate::algorithms::generate::generate_multi_prime_key_with_exp;
 use crate::dummy_rng::DummyRng;
 use crate::errors::{Error, Result};
 use crate::internals;
@@ -242,9 +242,13 @@ impl PublicKeyParts for RsaPrivateKey {
 }
 
 impl RsaPrivateKey {
+    /// Default exponent for RSA keys.
+    const EXP: u64 = 65537;
+
     /// Generate a new Rsa key pair of the given bit size using the passed in `rng`.
     pub fn new<R: CryptoRngCore + ?Sized>(rng: &mut R, bit_size: usize) -> Result<RsaPrivateKey> {
-        generate_multi_prime_key(rng, 2, bit_size)
+        let exp = BigUint::from_u64(Self::EXP).expect("invalid static exponent");
+        Self::new_with_exp(rng, bit_size, &exp)
     }
 
     /// Generate a new RSA key pair of the given bit size and the public exponent
@@ -256,7 +260,8 @@ impl RsaPrivateKey {
         bit_size: usize,
         exp: &BigUint,
     ) -> Result<RsaPrivateKey> {
-        generate_multi_prime_key_with_exp(rng, 2, bit_size, exp)
+        let components = generate_multi_prime_key_with_exp(rng, 2, bit_size, exp)?;
+        RsaPrivateKey::from_components(components.n, components.e, components.d, components.primes)
     }
 
     /// Constructs an RSA key pair from the individual components.
@@ -543,13 +548,18 @@ mod tests {
             #[test]
             fn $name() {
                 let mut rng = ChaCha8Rng::from_seed([42; 32]);
+                let exp = BigUint::from_u64(RsaPrivateKey::EXP).expect("invalid static exponent");
 
                 for _ in 0..10 {
-                    let private_key = if $multi == 2 {
-                        RsaPrivateKey::new(&mut rng, $size).expect("failed to generate key")
-                    } else {
-                        generate_multi_prime_key(&mut rng, $multi, $size).unwrap()
-                    };
+                    let components =
+                        generate_multi_prime_key_with_exp(&mut rng, $multi, $size, &exp).unwrap();
+                    let private_key = RsaPrivateKey::from_components(
+                        components.n,
+                        components.e,
+                        components.d,
+                        components.primes,
+                    )
+                    .unwrap();
                     assert_eq!(private_key.n().bits(), $size);
 
                     test_key_basics(&private_key);
@@ -568,17 +578,6 @@ mod tests {
     key_generation!(key_generation_multi_5_64, 5, 64);
     key_generation!(key_generation_multi_8_576, 8, 576);
     key_generation!(key_generation_multi_16_1024, 16, 1024);
-
-    #[test]
-    fn test_impossible_keys() {
-        let mut rng = ChaCha8Rng::from_seed([42; 32]);
-        for i in 0..32 {
-            let _ = RsaPrivateKey::new(&mut rng, i).is_err();
-            let _ = generate_multi_prime_key(&mut rng, 3, i);
-            let _ = generate_multi_prime_key(&mut rng, 4, i);
-            let _ = generate_multi_prime_key(&mut rng, 5, i);
-        }
-    }
 
     #[test]
     fn test_negative_decryption_value() {
